assignment

+++ src/Controller/productController.js
@@ -1,28 +1,31 @@
 import Product from "../Model/product.js";
-import Joi from "Joi";
+import Category from "../Model/category.js";
+import Joi from "joi";
+import { UpdateSchema } from "../validate/product.js";
 
 const productSchema = Joi.object({
   name: Joi.string().required(),
   price: Joi.number().required(),
   description: Joi.string().required(),
-  status: Joi.boolean().required(),
+  image: Joi.array().required(),
+  brand: Joi.string().required(),
+  status: Joi.boolean(),
   quality: Joi.number().required(),
+  categoryId: Joi.string().required(),
 });
 
 export const create = async (req, res) => {
   const data = req.body;
   try {
-    const { error } = productSchema.validate(data);
+    const { error } = productSchema.validate(req.body);
     if (!error) {
-      const newProduct = await Product({
-        name: req.body.name,
-        price: req.body.price,
-        description: req.body.description,
-        status: req.body.status,
-        quality: req.body.quality,
+      const newProduct = await Product.create(req.body);
+      await Category.findByIdAndUpdate(newProduct.categoryId, {
+        $addToSet: {
+          products: newProduct._id,
+        },
       });
-      await newProduct.save();
-      return res.status(200).json({ message: "tao thanh cong", newProduct });
+      return res.status(200).json({ message: "Tao thanh cong", newProduct });
     } else {
       return res.status(400).json({ message: error.message });
     }
@@ -30,14 +33,41 @@ export const create = async (req, res) => {
     return res.status(500).json({ message: error });
   }
 };
+
 export const getall = async (req, res) => {
+  const {
+    page = 1,
+    limit = 8,
+    sort = "createdAt",
+    order = "desc",
+    brand = "",
+    status = true,
+  } = req.query;
+  const options = {
+    page: page,
+    limit: limit,
+    sort: {
+      [sort]: order === "desc" ? -1 : 1,
+    },
+    populate: "categoryId",
+    collation: {
+      locale: "en",
+    },
+  };
   try {
-    const products = await Product.find({});
-    return res
-      .status(200)
-      .json({ message: "lay san pham thanh cong", products });
+    if (brand) {
+      const products = await Product.paginate({ brand, status }, options);
+      return res
+        .status(200)
+        .json({ message: "lay san pham thanh cong", products });
+    } else {
+      const products = await Product.paginate({ status }, options);
+      return res
+        .status(200)
+        .json({ message: "lay san pham thanh cong", products });
+    }
   } catch (error) {
-    return res.status(500).json({ message: error });
+    return res.status(500).json({ message: error.message });
   }
 };
 
@@ -57,8 +87,19 @@ export const getOne = async (req, res) => {
 export const update = async (req, res) => {
   const data = req.body;
   try {
-    const { error } = productSchema.validate(data);
+    const { error } = UpdateSchema.validate(data);
     if (!error) {
+      const oldProduct = await Product.findById(req.params.id);
+      await Category.findByIdAndUpdate(oldProduct.categoryId, {
+        $pull: {
+          products: oldProduct._id,
+        },
+      });
+      await Category.findByIdAndUpdate(data.categoryId, {
+        $addToSet: {
+          products: data._id,
+        },
+      });
       const updateProduct = await Product.findByIdAndUpdate(
         req.params.id,
         { $set: data },
@@ -74,14 +115,80 @@ export const update = async (req, res) => {
     return res.status(500).json({ message: error });
   }
 };
+
 export const remove = async (req, res) => {
   try {
+    const status = req.query.status;
     const id = req.params.id;
-    const removeProducts = await Product.findByIdAndDelete(id);
-    return res
-      .status(200)
-      .json({ message: "xoa san pham thanh cong", removeProducts });
+    if (status) {
+      const changeStatusProduct = await Product.findByIdAndUpdate(
+        id,
+        {
+          $set: { status: status },
+        },
+        { new: true }
+      );
+      return res.status(200).json({
+        message: "sản phẩm đã chuyển đến thùng rác ",
+        changeStatusProduct,
+      });
+    } else {
+      const removeProducts = await Product.findByIdAndDelete(id);
+      return res
+        .status(200)
+        .json({ message: "Đã xoá vĩnh viễn sản phẩm", removeProducts });
+    }
   } catch (error) {
     return res.status(500).json({ message: error });
+  }
+};
+
+export const restore = async (req, res) => {
+  try {
+    const id = req.params.id;
+    const changeStatusProduct = await Product.findByIdAndUpdate(
+      id,
+      {
+        $set: { status: true },
+      },
+      { new: true }
+    );
+    return res.status(200).json({
+      message: "sản phẩm đã được khôi phục",
+      changeStatusProduct,
+    });
+  } catch (error) {
+    return res.status(500).json({ message: error });
+  }
+};
+
+export const getRelase = async (req, res) => {
+  const {
+    page = 1,
+    limit = 8,
+    sort = "createdAt",
+    order = "desc",
+    status = true,
+  } = req.query;
+  const options = {
+    page: page,
+    limit: limit,
+    sort: {
+      [sort]: order === "desc" ? -1 : 1,
+    },
+    populate: "categoryId",
+    collation: {
+      locale: "en",
+    },
+  };
+  try {
+    const categoryId = req.params.categoryId;
+    console.log(categoryId);
+    const products = await Product.paginate({ status, categoryId }, options);
+    return res
+      .status(200)
+      .json({ message: "lay san pham thanh cong", products });
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
   }
 };
